CASS-56 | Remove behaviour from api classes
