+++ src/main/java/org/mule/modules/cassandradb/api/ReplicationStrategy.java
@@ -45,53 +45,4 @@ public enum ReplicationStrategy {
     public String getStrategyClass() {
         return this.strategy;
     }
-
-    public static ReplicationStrategy lookup(String inputToMatch) {
-        if (StringUtils.isNotBlank(inputToMatch)) {
-            for (ReplicationStrategy enumItem : ReplicationStrategy.values()) {
-                if (enumItem.getStrategyClass().equalsIgnoreCase(inputToMatch)) {
-                    return enumItem;
-                }
-            }
-        }
-        return null;
-    }
-
-    //based on http://docs.datastax.com/en/cql/3.1/cql/cql_reference/create_keyspace_r.html rules
-    public static Map<String, Object> buildReplicationStrategy(CreateKeyspaceInput input) {
-        LinkedHashMap<String, Object> replicationStrategy = new LinkedHashMap<String, Object>();
-
-        //set the strategy class only if exists in the input && a valid one is provided
-        if (StringUtils.isNotBlank(input.getReplicationStrategyClass())) {
-            ReplicationStrategy strategy = ReplicationStrategy.lookup(input.getReplicationStrategyClass());
-            //setting 'replication_factor' || 'first_data_center' || 'next data center' is pointless if no class is provided
-            if (strategy != null) {
-                replicationStrategy.put(Constants.CLASS, input.getReplicationStrategyClass());
-                //'replication_factor' required if class is SimpleStrategy; otherwise, not used
-                if (strategy.equals(ReplicationStrategy.SIMPLE)) {
-                    replicationStrategy.put(Constants.REPLICATION_FACTOR, input.getReplicationFactor());
-                }
-                if (input.getFirstDataCenter() != null) {
-                    if (StringUtils.isNotBlank(input.getFirstDataCenter().getName())) {
-                        replicationStrategy.put(input.getFirstDataCenter().getName(), input.getFirstDataCenter().getValue());
-                    }
-                }
-                if (input.getNextDataCenter() != null) {
-                    if (StringUtils.isNotBlank(input.getNextDataCenter().getName())) {
-                        replicationStrategy.put(input.getNextDataCenter().getName(), input.getNextDataCenter().getValue());
-                    }
-                }
-            }
-        } else {
-            return buildDefaultReplicationStrategy();
-        }
-        return replicationStrategy;
-    }
-
-    public static Map<String, Object> buildDefaultReplicationStrategy() {
-        LinkedHashMap<String, Object> replicationStrategyMap = new LinkedHashMap<String, Object>();
-        replicationStrategyMap.put(Constants.CLASS, ReplicationStrategy.SIMPLE.getStrategyClass());
-        replicationStrategyMap.put(Constants.REPLICATION_FACTOR, Constants.DEFAULT_REPLICATION_FACTOR);
-        return replicationStrategyMap;
-    }
 }
